Modifications to allow adjustable symbol size.
Properties is now an instantiation of a template, that allows
us to ajust the size of the symbols in the alphabet we are encoding.
This change is primarily for testing. Many places have been changed
to deal with this. Most common is an extra argument passed into
various functions, to avoid having to turn those functions into
templates.