Reorganize to handle EventStore lifetime.
The EventStore will be in charge of managing the thread that runs
the function that consumes events put onto the shared event queue.
The plan is to have the destruction of the EventStore make the call
to 'join' on that thread, so that the destructor will not exit
until the thread is done working.