Add start of testing code to SimpleQueueReader.
SimpleQueueReader now always expects 100 events to be processed,
and prints a message to stderr if it finds otherwise. The 100
should be changed to be a value from a configuration parameter.
We also need a way for the program to indicate it hasn't read the
expected number of elements.

Perhaps an expected number of 0 or the maximum size_t should indicate
that we want NO limit testing to be done.