Configuration behavior for RawInput upon timeout.
One can now configure RawInput's response to a timeout while waiting
to read from the shared event queue. The default is to keep trying
to read; this is the result from setting resume_after_timeout to
'true'. If set to false, a timeout is handled the same was as an
end-of-data.
If the call to read the shared event queue times out,