Unlike previous commit on this branch, eventbuilder will now work with ds50daq
artdaq's eventbuilder now, like in ds50daq's eventbuilder, will
declare a communicator group of the eventbuilder processes at the
level of the eventbuilder.cc file, rather than inside the EventBuilder
class object. Also, during initialization in EventBuilderApp, the
pointer to the EventBuilder object is only reset if it's currently
null; this is the way things are done in ds50daq code.