Redesigned FragmentGenerator to contain virtual "final" state transition functions calling overridable private virtual functions
After a discussion with Kurt and Ron this morning (Friday, 12/6/13),
and with input from Marc Paterno, we decided to rework
FragmentGenerator to contain state transition functions (stop, start,
pause, resume, as well as report) NOT meant to be overriden in derived
classes, but rather containing code intended to be used by all classes
derived from FragmentGenerator as well as calls to private virtual
functions which derived classes can override so as to provide
experiment-specific functionality. Aside from the formal reworking of
FragmentGenerator, some of the functionality implemented in
DS50FragmentGenerator's state transition functions has now been moved
into FragmentGenerator (its parent class) so that, which DS50's code
effectively does the same thing it did before, some of its
implementation is now available to all users of FragmentGenerator.

Some things to keep in mind: it's not yet final what functionality
should be added in from DS50FragmentGenerator (in particular, the
entire implementation of getNext_ is in DS50FragmentGenerator), and
the mutex_ object has been declared protected for now; while
convenient, this should be revisited as it's not best practice to use
protected data.