Change names of executables and their source files
In response to Chris Green's pointing out that case-insensitivity on
certain platforms could cause problems given that we had, e.g., a file
named eventbuilder.cc and a file named EventBuilder.cc, we've renamed
the source files and the class they're named after. To give, as an
example, eventbuilder:

eventbuilder(.*) : EventBuilderMain(.*)
EventBuilderApp.* : Same as before
EventBuilder.cc : EventBuilderCore.cc

Please also note that "FragmentReceiver" has been renamed
"BoardReaderCore", in keeping with the naming conventions for the
eventbuilder and aggregator apps.