MPISentry now sets up local group in its constructor and throws in case of error
After a discussion with Kurt today, we decided that rather than let
the process group (eventbuilder, aggregator, boardreader) set up be
part of a separate function call -- MPISentry::create_local_group --
it should instead be done in the constructor after the world group
gets set up. If anything goes wrong, an exception is
thrown. Consequently, the MPISentry object is now constructed in a
try-catch block within the main() function in the executables, and if
something goes wrong, the constructor throws a cet::exception object
with an error message, which is then re-thrown by the executable. If
things go off without a hitch, the local process group is returned to
the user via a pass-by-reference in the constructor.