Created the MetricPlugin interface. Added factory class makeMetricPlugin. Included this MetricPlugin factory in StatisticsHelper so that whenever a metric is passed to StatisticsHelper, it also sends it along to any metric plugins that have been defined in the fhicl "metrics" block.
Syntax of metrics block:

metrics : {
  names : [ file, ganglia ]
  file : {
    fileName : "MetricOutput.log"
    level : 5
  }
  ganglia : {
    configFile: "/etc/ganglia/gmond.conf"
    groupName: "BoardReaderMetrics"
    level : 3
  }
}

Notes: The "names" array MUST contain the names of the modules so that
the MetricPlugin factory can find them. The config blocks for the modules
MUST ALSO be prefaced with the module name.

Absence of a "metrics" config block is not an error and will simply cause
an informative message to be printed.